#75 - Repository interface extracted
