+++ src/test/java/com/artipie/composer/http/PhpComposerTest.java
@@ -28,7 +28,7 @@ import com.artipie.asto.Storage;
 import com.artipie.asto.blocking.BlockingStorage;
 import com.artipie.asto.memory.InMemoryStorage;
 import com.artipie.composer.AllPackages;
-import com.artipie.composer.Repository;
+import com.artipie.composer.AstoRepository;
 import com.artipie.http.Response;
 import com.artipie.http.hm.RsHasBody;
 import com.artipie.http.hm.RsHasStatus;
@@ -75,7 +75,7 @@ class PhpComposerTest {
     @BeforeEach
     void init() {
         this.storage = new InMemoryStorage();
-        this.php = new PhpComposer(new Repository(this.storage));
+        this.php = new PhpComposer(new AstoRepository(this.storage));
     }
 
     @Test
